# Conflicts:
#	Java/JDI/jdi-cucumber-test-generator/src/main/java/com/epam/test_generator/controllers/CaseController.java
#	Java/JDI/jdi-cucumber-test-generator/src/main/java/com/epam/test_generator/controllers/StepController.java
#	Java/JDI/jdi-cucumber-test-generator/src/main/java/com/epam/test_generator/controllers/SuitController.java
#	Java/JDI/jdi-cucumber-test-generator/src/main/webapp/WEB-INF/static/views/newSuits.html

+++ Java/JDI/jdi-cucumber-test-generator/src/main/java/com/epam/test_generator/controllers/SuitController.java
@@ -25,56 +25,79 @@ public class SuitController {
 
     @RequestMapping(value = "/")
     public String getMainPage() {
+	@RequestMapping(value = "/")
+    public String getMainPage() {
         return "/WEB-INF/static/views/newSuits";
     }
 
-    @RequestMapping(value = "/getAllSuits", method = RequestMethod.GET, produces = "application/json")
+    @RequestMapping(value = "/suits", method = RequestMethod.GET, produces = "application/json")
     @ResponseBody
     public ResponseEntity<List<SuitDTO>> getSuits() {
-
-        return new ResponseEntity<>(suitService.getSuits(),HttpStatus.OK);
+		return new ResponseEntity<>(suitService.getSuits(), HttpStatus.OK);
     }
 
-    @RequestMapping(value = "/getSuit/{id}", method = RequestMethod.GET)
+    @RequestMapping(value = "/suit/{suitId}", method = RequestMethod.GET)
     @ResponseBody
-    public ResponseEntity<SuitDTO> getSuit(@PathVariable("id") long id){
+    public ResponseEntity<SuitDTO> getSuit(@PathVariable("suitId") long id){
+        SuitDTO suitDTO = suitService.getSuit(id);
 
-        return new ResponseEntity<>(suitService.getSuit(id), HttpStatus.OK);
+		if (suitDTO != null) {
+			return new ResponseEntity<>(suitDTO, HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    @RequestMapping(value="/updateSuit", method = RequestMethod.PATCH, consumes = "application/json")
-    public ResponseEntity<Void> editSuit(@RequestBody SuitDTO suit){
-        suitService.updateSuit(suit);
-        return new ResponseEntity<>(HttpStatus.OK);
+    @RequestMapping(value="/suit/{suitId}", method = RequestMethod.PUT, consumes = "application/json")
+    public ResponseEntity<Void> editSuit( @PathVariable("suitId") long id, @RequestBody SuitDTO suitDTO){
+        if (isNameValid(suitDTO.getName()) && isPriorityValid(suitDTO.getPriority()) &&
+                isDescriptionValid(suitDTO.getDescription()) && isTagsValid(suitDTO.getTags())) {
+            suitService.updateSuit(suitDTO);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.UNPROCESSABLE_ENTITY);
     }
 
-    @RequestMapping(value = "/removeSuit/{id}", method = RequestMethod.DELETE)
-    public ResponseEntity<Void> removeSuit(@PathVariable("id") long id){
-        suitService.removeSuit(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    @RequestMapping(value = "/suit/{suitId}", method = RequestMethod.DELETE)
+    public ResponseEntity<Void> removeSuit(@PathVariable("suitId") long id){
+        SuitDTO suitDTO = suitService.getSuit(id);
+
+        if (suitDTO != null) {
+            suitService.removeSuit(id);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
-    @RequestMapping(value="/addSuit", method = RequestMethod.PUT, consumes = "application/json", produces = "application/json")
+    @RequestMapping(value="/suit", method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
     @ResponseBody
-    public ResponseEntity<SuitDTO> addSuit(@RequestBody SuitDTO suit) {
-        return new ResponseEntity<>(suitService.addSuit(suit), HttpStatus.OK);
+    public ResponseEntity<SuitDTO> addSuit(@RequestBody SuitDTO suitDTO) {
+        if (isNameValid(suitDTO.getName()) && isPriorityValid(suitDTO.getPriority()) &&
+                isDescriptionValid(suitDTO.getDescription()) && isTagsValid(suitDTO.getTags())) {
+			return new ResponseEntity<>(suitService.addSuit(suitDTO), HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(suitDTO, HttpStatus.UNPROCESSABLE_ENTITY);
     }
 
-    @RequestMapping(value = "/downloadFeatureFile", method = RequestMethod.GET)
-    @ResponseBody
-    public String downloadFile(@RequestParam(name = "suitId") Long suitId,
-                             @RequestParam(name = "caseIds") List<Long> caseIds,
-                             HttpServletResponse response) throws IOException {
-
-        String mimeType = "application/octet-stream";
-        System.out.println("mimetype : " + mimeType);
-        response.setContentType(mimeType);
-        response.setHeader("Content-Disposition", "inline; filename=\"" + "File.feature" + "\"");
-
-        return  suitService.generateStream(suitId, caseIds);
-//        ByteArrayInputStream inputStream = suitService.generateStream(suitId, caseIds);
-//        response.setContentLength(inputStream.available());
-//        FileCopyUtils.copy(inputStream, response.getOutputStream());
-//        inputStream.close();
+    private boolean isPriorityValid(Integer priority) {
+        return priority != null && priority >= 1 && priority <= 5;
     }
+
+    private boolean isNameValid(String name) {
+        return name != null && name.length() >= 1 && name.length() <= 255;
+    }
+
+    private boolean isDescriptionValid(String description) {
+        return description == null || description.length() <= 255;
+    }
+
+    private boolean isTagsValid(String tags) {
+        return tags == null || tags.length() <= 255;
+    }
+
 }
